requirements.txt
numpy
Pillow
scikit-image
pybrisque
libsvm
openai>=1.0.0
termcolor
python-dotenv
scipy
opencv-python
httpx
